requirements.txt
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
pandas==2.1.3
notion-client==2.2.1
python-dotenv==1.0.0
schedule==1.2.1 